+++ script/provider/completion.lua
@@ -3,8 +3,8 @@ local proto = require 'proto'
 local isEnable = false
 
 local function allWords()
-    local str = [[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789.:('"[,#*@|=- ]]
-    local list = {}
+    local str = [[abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789.:('"[,#*@|=-{ ]]
+    local list = {'\n', '\t'}
     for c in str:gmatch '.' do
         list[#list+1] = c
     end
@@ -48,6 +48,7 @@ local function disable()
 end
 
 return {
-    enable = enable,
-    disable = disable,
+    enable   = enable,
+    disable  = disable,
+    allWords = allWords,
 }